home/ubuntu/inventory_management/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
psycopg2-binary>=2.9,<3.0
redis>=4.0,<5.0
kafka-python>=2.0,<3.0
ultralytics>=8.0,<9.0
opencv-python-headless>=4.5,<5.0
numpy>=1.21,<2.0
djangorestframework-simplejwt>=4.7,<5.0
django-cors-headers>=3.7,<4.0
django-redis>=5.0,<6.0
fastapi>=0.68,<1.0
uvicorn>=0.15,<1.0
passlib[bcrypt]>=1.7,<2.0
pyjwt>=2.1,<3.0
python-multipart>=0.0.5,<0.1
